+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [1.0.2] - 23-02-2025
+
+### Improvements
+- Added error checking for when the Zabbix Agent has no installation candidate
+- Added proper detection of the latest stable Zabbix version
+
 ## [1.0.1] - 23-02-2025
 
 ### Improvements